// src/api/kochservice.schemas.ts
/**
 * Generated by orval v7.20.0 🍺
 * Do not edit manually.
 * kochservice
 * OpenAPI spec version: 0.1.0
 */
export interface CreateIngredientRequest {
  name: string
}

export interface CreateRecipeRequest {
  ingredients: RecipeIngredientDto[]
  instructions: string[]
  name: string
  tag_ids: string[]
}

export interface CreateTagRequest {
  name: string
}

export interface IngredientResponse {
  id: string
  name: string
}

export interface IngredientsResponse {
  ingredients: IngredientResponse[]
}

export interface RecipeIngredientDto {
  amount: number
  ingredient_id: string
  measurement: string
}

export interface RecipeIngredientResponse {
  amount: number
  id: string
  ingredient_id: string
  ingredient_name: string
  measurement: string
}

export interface RecipeResponse {
  id: string
  ingredients: RecipeIngredientResponse[]
  instructions: string[]
  name: string
  tags: TagResponse[]
}

export interface RecipesResponse {
  recipes: RecipeResponse[]
  /** @minimum 0 */
  total_pages: number
}

export interface TagResponse {
  id: string
  name: string
}

export interface TagsResponse {
  tags: TagResponse[]
}

export interface UpdateIngredientRequest {
  name: string
}

export interface UpdateRecipeRequest {
  ingredients: RecipeIngredientDto[]
  instructions: string[]
  name: string
  tag_ids: string[]
}

export interface UpdateTagRequest {
  name: string
}

export type GetRecipesParams = {
  /**
   * Page number (0-indexed)
   * @minimum 0
   */
  page: number
  /**
   * Items per page
   * @minimum 0
   */
  page_size: number
}